fix: default tab query
